MSVC reports "SVGPropertyAnimator.h(94): error C2839: invalid return type 'T *' for overloaded 'operator ->'" with /std:c++20 https://bugs.webkit.org/show_bug.cgi?id=234546 <rdar://problem/86757805>
Reviewed by Darin Adler.

r287300 was a wrong fix that used a raw pointer to work around an
MSVC's RefPtr compilation error. We should use smart pointers for
local variables.

This seems a MSVC bug that the class template argument deduction
doesn't work in this case. Explicitly specifying a class template
argument is a workaround.

* svg/properties/SVGPropertyAnimator.h:
(WebCore::SVGPropertyAnimator::computeInheritedCSSPropertyValue const):
Changed the type of the local variable 'parent' from a raw pointer
to RefPtr<Element>.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288913 268f45cc-cd09-0410-ab3c-d52691b4dbfc