scroll-margin-top doesn't work on inline elements https://bugs.webkit.org/show_bug.cgi?id=235933 <rdar://87983307>
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

* web-platform-tests/css/css-scroll-snap/scroll-target-margin-006-expected.txt: Added.
* web-platform-tests/css/css-scroll-snap/scroll-target-margin-006.html: Added.

Source/WebCore:

Test: imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-target-margin-005.html

Move the implementation of absoluteAnchorRectWithScrollMargin which actually processes
the CSS scroll-margin property to RenderElement. This allows it to be used for inline
as well as block elements. The specification says it should apply to both. The box
sizes passed in here do not matter too much, because the Length should never be a
percentage.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::absoluteAnchorRectWithScrollMargin const): Deleted.
* rendering/RenderBox.h:
* rendering/RenderElement.cpp:
(WebCore::RenderElement::absoluteAnchorRectWithScrollMargin const):
* rendering/RenderElement.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288947 268f45cc-cd09-0410-ab3c-d52691b4dbfc