LibWebRTCCodecs should not need to create IOSurfaces https://bugs.webkit.org/show_bug.cgi?id=235951 <rdar://problem/88326654>
Reviewed by Eric Carlson.

Source/ThirdParty/libwebrtc:

Introduce helper routines to copy a webrtc VideoFrame into a memory buffer and get a CVPixelBuffer from a webrtc VideoFrame.

* Configurations/libwebrtc.iOS.exp:
* Configurations/libwebrtc.iOSsim.exp:
* Configurations/libwebrtc.mac.exp:
* Source/webrtc/sdk/WebKit/WebKitUtilities.h:
* Source/webrtc/sdk/WebKit/WebKitUtilities.mm:

Source/WebCore:

Introduce SharedVideoFrameInfo as a helper class to serialize a CVPixelBuffer into memory and create a CVPixelBuffer from reading memory.

Covered by existing webrtc tests.

* SourcesCocoa.txt:
* WebCore.xcodeproj/project.pbxproj:
* platform/cocoa/SharedVideoFrameInfo.h: Added.
* platform/cocoa/SharedVideoFrameInfo.mm: Added.
* platform/cocoa/CoreVideoSoftLink.cpp:
* platform/cocoa/CoreVideoSoftLink.h:
* platform/graphics/RemoteVideoSample.cpp:
* platform/graphics/RemoteVideoSample.h:

Source/WebKit:

Introduce SharedVideoFrameWriter and SharedVideoFrameReader to support shared memory sending of CVPixelBuffers that are non IOSurface backed.
To do so, we make use of SharedVideoFrameInfo from WebCore.
What we do is:
- When getting a frame, allocate a shared memory that is big enough to copy its data and info in it. Create a semaphore to sync reader and writer.
- If we allocate shared memory/semphore, send them over IPC to the SharedVideoFrameReader.
- Wait for semaphore to tell us reader is fine for writing.
- Write SharedVideoFrameInfo into the shared memory, then copy the raw bytes into the shared memory.
- Send through IPC the video frame, no MacSendRight means to read the shared memory.
On reader side, for every frame that does not have an IOSurface, we read the shared memory, create an IOSurface from it and notify WebProcess that it can write the next frame.

In memory buffers are used for black frames at least.
webrtc/video-mute.html covers the patch.

* GPUProcess/webrtc/LibWebRTCCodecsProxy.h:
* GPUProcess/webrtc/LibWebRTCCodecsProxy.messages.in:
* GPUProcess/webrtc/LibWebRTCCodecsProxy.mm:
* SourcesCocoa.txt:
* WebKit.xcodeproj/project.pbxproj:
* WebProcess/GPU/webrtc/LibWebRTCCodecs.cpp:
* WebProcess/GPU/webrtc/LibWebRTCCodecs.h:
* WebProcess/GPU/webrtc/SharedVideoFrame.cpp: Added.
* WebProcess/GPU/webrtc/SharedVideoFrame.h: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288951 268f45cc-cd09-0410-ab3c-d52691b4dbfc