REGRESSION(r288865): SourceImage should never sink its ImageBuffer to a NativeImage https://bugs.webkit.org/show_bug.cgi?id=236005
Reviewed by Simon Fraser.

Source/WebCore:

The r288865 changes SourceImage::nativeImage() such that if the image
variant has an ImageBuffer, it will sink this ImageBuffer into a NativeImage
and return this NativeImage to the caller. This was incorrect change
because this would invalidate the ImageBufferBackend of the SourceImage.

The fix is to copy the ImageBuffer to a NativeImage (DontCopyBackingStore)
and store it into another image variant. This will keep the ImageBuffer
valid and will prevent subsequent conversion if this function is called
multiple times.

Similar changes should be applied to SourceImage::imageBuffer().

Test: svg/custom/pattern-multiple-referencing.html

* platform/graphics/SourceImage.cpp:
(WebCore::nativeImageOf):
(WebCore::SourceImage::nativeImageIfExists const):
(WebCore::SourceImage::nativeImage const):
(WebCore::imageBufferOf):
(WebCore::SourceImage::imageBufferIfExists const):
(WebCore::SourceImage::imageBuffer const):
* platform/graphics/SourceImage.h:

LayoutTests:

* svg/custom/pattern-multiple-referencing-expected.txt: Added.
* svg/custom/pattern-multiple-referencing.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288977 268f45cc-cd09-0410-ab3c-d52691b4dbfc