Crash in AXIsolatedObject::associatedAXObject. https://bugs.webkit.org/show_bug.cgi?id=236053 <rdar://problem/88413232>
Reviewed by Darin Adler.

The crash occurs because we are using an axObject pointer after we call
updateBackingStore() which may destroy the object pointed to by axObject.
This was added as a fix for
https://bugs.webkit.org/show_bug.cgi?id=220446
So I will submit a follow up patch with the correct fix for that bug as well.

* accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::associatedAXObject const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289044 268f45cc-cd09-0410-ab3c-d52691b4dbfc