AX: When updating the isolated tree with a remove and add of the same object, the wrapper is not re-attached https://bugs.webkit.org/show_bug.cgi?id=236273
Reviewed by Andres Gonzalez.

In AXIsolatedTree::applyPendingChanges, we process node and subtree
removals before tree additions. When doing these removals, we detach
the object and wrapper from each other. This is fine unless that same
object is slated to be added again (e.g. it was moved in the DOM), as
we don't always re-attach a wrapper. [WebAccessibilityObjectWrapperBase axBackingObject]
will always return nil for objects in this state.

In this patch, when processing m_pendingAppends, we now attach a
wrapper to any object that has become `isDetached` (i.e. the object
has no wrapper, or the object has a wrapper but the wrapper does not
point at the object).

Fixes ~15 tests in isolated tree mode.

* accessibility/isolatedtree/AXIsolatedObject.cpp:
* accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::applyPendingChanges):
* accessibility/isolatedtree/mac/AXIsolatedObjectMac.mm:
(WebCore::AXIsolatedObject::isDetached const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289368 268f45cc-cd09-0410-ab3c-d52691b4dbfc