Conversion to a color space with a smaller gamut should perform gamut mapping https://bugs.webkit.org/show_bug.cgi?id=236200
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

Update tests to include more examples that force gamut mapping and update the results
of a few existing ones to account for the new algorithm.

* web-platform-tests/css/css-color/parsing/color-mix-computed-expected.txt:
* web-platform-tests/css/css-color/parsing/color-mix-computed.html:
* web-platform-tests/css/css-color/parsing/color-mix-valid-expected.txt:
* web-platform-tests/css/css-color/parsing/color-mix-valid.html:
* web-platform-tests/css/css-color/parsing/relative-color-computed-expected.txt:
* web-platform-tests/css/css-color/parsing/relative-color-computed.html:
* web-platform-tests/css/css-color/parsing/relative-color-valid-expected.txt:
* web-platform-tests/css/css-color/parsing/relative-color-valid.html:

Source/WebCore:

CSS Color now defines that gamut mapping should happen when converting to an RGB color space
with bounded gamut that is smaller than the origin color's color space. Specifically, it
specifies the use of a new "CSS gamut mapping algorithm" https://drafts.csswg.org/css-color/#css-gamut-mapping
which implements a relative colorimetric intent mapping with colors inside the destination
gamut unchanged.

The previous behavior we implemented was to clip out of gamut colors to the gamut (e.g.
color(srgb 1.4 -0.2 .5) would become color(srgb 1 0 .5)) which can lead to very odd results.

To keep things simple, the gamut mapping has been incorporated directly into the main color
conversion pipeline, replacing the call to makeFromComponentsClampingExceptAlpha in toBounded()
with one implements the new algorithm. To accomadate this, the pipeline was modified so that
the matrix conversions now only happen to/from extended-linear and XYZ color types, rather than
the old behavior which allowed shortcuting from bounded-linear, which was ok because the clip
would result in the same value. Since conversion from bounded-linear to extended-linear is free
(just a type change, no conversion), the only additional cost here is the gamut mapping.

The implementation of the CSS gamut mapping algorithm itself is a naive iterative bisection
implementation based on the psuedo-code from the spec. As an optimization in the future, we should
consider implementing an analytic solution, which will be more complicated, but likely faster.

Currently, the only place where the effect of this new gamut mapping behavior will be visible is
via functions that explicitly convert to a bounded color space, specifically the CSS color-mix()
function and relative color syntax. Ultimately, we should also be using this to map CSS colors
to the output color space of the screen / HTML canvas, but finding the right place to do that
is not a part of this initial change.

* platform/graphics/ColorConversion.cpp:
(WebCore::ColorConversion<Lab<float>, XYZA<float, WhitePoint::D50>>::convert):
(WebCore::ColorConversion<OKLab<float>, XYZA<float, WhitePoint::D65>>::convert):
Add call to makeFromComponentsClampingExceptAlpha at the end of conversion to Lab/OKLab
as the conversion functions can produce lightness values just ever so less than 0.
In practice, only the lightness value is bounded, so it is the only thing that actually
gets clamped thanks to the constexpr nature of clampedComponent().

* platform/graphics/ColorConversion.h:
(WebCore::computeDeltaEOK):
Add implementation of the deltaEOK algorithm.

(WebCore::ClipGamutMapping::mapToBoundedGamut):
Add implementation of the degenerate gamut mapping algorithm.

(WebCore::CSSGamutMapping::mapToBoundedGamut):
Add implementation of the CSS gamut mapping algorithm.

(WebCore::ColorConversion::convert):
Add a new step to the main pipeline that explicitly handles bounds conversions to/from
extended-linear so that the matrix conversions can rely on always working in extended space.

(WebCore::ColorConversion::toBounded):
Replace call to clamp with call out to gamut mapping algorithm.

(WebCore::ColorConversion::handleMatrixConversion):
Replace makeFromComponentsClampingExceptAlpha with makeFromComponents now that all conversions
are in the extended space and clamping is no-op.

* platform/graphics/ColorTypes.h:
(WebCore::inGamut):
(WebCore::colorIfInGamut):
(WebCore::clipToGamut):
Add some helpers to check/convert/clip to the bounded gamut of RGB types.

Tools:

* TestWebKitAPI/Tests/WebCore/ColorTests.cpp:
(TestWebKitAPI::TEST):
Update color conversion tests to account for gamut mapping and add a new test
that bypasses gamut mapping by converting to extended sRGB.

LayoutTests:

* platform/gtk/TestExpectations:
Add more tests that are only failing on the GTK bots due to a small floating point issues that needs further investigation.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289396 268f45cc-cd09-0410-ab3c-d52691b4dbfc