[web-animations] additive and accumulation interpolation does not work correctly with implicit 0% and 100% keyframes https://bugs.webkit.org/show_bug.cgi?id=236314
Reviewed by Dean Jackson.

LayoutTests/imported/w3c:

Add some new WPT tests and mark our PASS results for the "add" case (we still fail for "accumulate").

* web-platform-tests/web-animations/animation-model/combining-effects/effect-composition-expected.txt:
* web-platform-tests/web-animations/animation-model/combining-effects/effect-composition.html:

Source/WebCore:

We incorrectly handled implicit keyframes for the additive and accumulate cases.

The spec says that for implicit 0% and 100% keyframes, a keyframe should be generated with a "neutral" keyframe which,
when added or accumulated with another keyframe, would yield the same style as the keyframe it's composed with. We sort
of did the right thing by cloning the underlying style for those keyframes, but then we would blame them anyway in the
composition case, whereas we should just use the underlying style as-is.

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::setAnimatedPropertiesInStyle):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289454 268f45cc-cd09-0410-ab3c-d52691b4dbfc