[LFC][IFC] Vertical writing mode with RTL text content has incorrect advances https://bugs.webkit.org/show_bug.cgi?id=236345
Reviewed by Antti Koivisto.

contentRightInInlineDirectionVisualOrder is the visual offset of the adjoining runs ignoring the writing mode (horizontal vs. vertical).
It's visual in the inline direction sense (right to left vs. left to right), but it's not transformed in the context of writing mode, i.e. width and height are not flipped.

* layout/formattingContexts/inline/display/InlineDisplayContentBuilder.cpp:
(WebCore::Layout::InlineDisplayContentBuilder::processBidiContent):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289499 268f45cc-cd09-0410-ab3c-d52691b4dbfc