Fail synchronously when constructing a SharedWorker with an URL that is not same-origin https://bugs.webkit.org/show_bug.cgi?id=236419
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

* web-platform-tests/workers/constructors/SharedWorker/same-origin-expected.txt:
Rebaseline test that is now fully passing. I have verified that it is passing in both
Blink and Gecko too.

* web-platform-tests/workers/shared-worker-in-data-url-context.window-expected.txt:
Even though this looks like a regression, this actually aligns our behavior with both
Blink & Gecko (who also fail this check). Note that the load fails no matter what.
However, the test expects it to fail asynchronously instead of synchronously in this
case.

Source/WebCore:

Fail synchronously when constructing a SharedWorker with an URL that is not same-origin.
This aligns our behavior with Chrome and matches the language in the specification.

No new tests, rebaselined existing test.

* workers/shared/SharedWorker.cpp:
(WebCore::SharedWorker::create):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289532 268f45cc-cd09-0410-ab3c-d52691b4dbfc