AX: Log more information when the wrapper or the wrapper's backing object is detached https://bugs.webkit.org/show_bug.cgi?id=236381
Reviewed by Chris Fleizach.

Right now, we don't log any information when we detect a wrapper has
no object, or when it does have an object but it's detached from a
parent.

This patch logs the wrapper address and object in these scenarios, which
will help debug these issues in conjuction with other logs.

* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper logMissingBackingObject]):
Added.
(-[WebAccessibilityObjectWrapper accessibilityAttributeValue:]):
Use new logMissingBackingObject, and log object if it's detached from a parent.
(-[WebAccessibilityObjectWrapper _accessibilitySetValue:forAttribute:]):
Use new logMissingBackingObject.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289538 268f45cc-cd09-0410-ab3c-d52691b4dbfc